Added base for reversibility and some documentation.
- Reversibility is entirely handled by the superclass Process.java, because the reversal process is pretty much the same for every kind of process.

- Added complement label identification: SetUtil.getTauMatches

-- !Important! Explanation of the mechanics of current reversibility:

Whenever a process is acted on by a label, the process stores itself in a private variable along with a key generated from that label. That way, every process that has acted at least once will have the ability to just 'return' the variable stored in memory, along with its associated key. This is **NOT** how I want it to work in the future, and I don't think it's very refined. I am only adding it as a PoC right now.

Signed-off-by: GMX <pbrowning7@gmail.com>