- Change clone() method type
- Add more clone implementations

- Change test string matching format

I'm unsure of how to deal with previous lives. I decided that it was worth giving this method a little more work before changing to a different system altogether. Because of how the system works, a.b|c.d acting on a will first call the parent process (a.b|c.d), which will then call the child process (a.b). So, the question arises: Which one should remember the previous life (and thus hold the key)? Well, in this case it is easy to see that the child process should. But what are the rules? I'm not sure yet. It is possible that I may just give the parent process the key, but it's a difficult problem.

Signed-off-by: GMX <pbrowning7@gmail.com>