Tau fixes, implemented synchronization locking.
An issue I talked about previously with summation processes and tau synchronizations has been fixed by adding a protected synchronizationLock array of Labels to each label, indicating which labels they are forbidden from synchronizing with. Then, I added an override getActionableLabels method in the SummationProcess class which then loops through all actionable labels on the left, and adds a lock so they can synchronize with any on the right, and then vice versa. This is both an elegant and non-elegant solution at the same time, but it does work.

Signed-off-by: GMX <pbrowning7@gmail.com>