Fixed issue causing multiple instances of getActionableLabels() to be called.
Also fixed an issue causing nested concurrent processes to compound the tau nodes from child processes. So, 0|a|'a would have a tau node for the right child, but then the parent would, because it has the same actionable labels, duplicate the tau node causing multiple tau nodes to show up in the collection.

Performance should also be increased because of the fewer calls to getActionableLabels

Signed-off-by: GMX <pbrowning7@gmail.com>