One commit to rule them all >:)
After many headaches and posture readjustments I finally figured out a good way to set up annotation vs. removal. I did **A LOT** in this commit, but there will only be this one commit because the code was pretty much broken and not working the entire time I was working on it, only working at the very very end. I'll list some important changes introduced:

ActionPrefixProcess now overrides canAct() to represent that it is able to act on parts of a tau node.

Tau nodes are now actionable

There is now a new configuration option that allows users to choose how they want to see summation annotation (or lack thereof). To briefly sum it up, there are 3 ways it can be displayed. The example I'll use is (a + b|c) acting on b.
Mode 0: Implicit annotation: It would be displayed at (a + [b]|c). This mode is meant for only the true hardcore CCSers out there, who can determine whether or not a process is actionable at a single glance. Realistically though, this is a very lightweight solution that can work in all situations. Because one side of the summation process is annotated, that implies that it has been acted on and thus the other side cannot be acted on

Mode 2: Annotation (easy mode)
This mode will annotate the process to look like
([b]{a} + [b]|c). This shows that the left side (a) is locked behind a key ([b]). This is sort of like cheating but useful during debugging

Mode 3: Ghosting (IDGAF about CCS mode). This mode will completely hide the side of the process that cannot be acted on. It would look like this: ([b]|c). From looking at this process alone, there is no way to tell what the process may be after rewinding on b, but if that is not needed then this is a much nicer way to view processes.

ActionPrefixProcesses can now ghost and utilize ghost keys (long story short, I got creative. If anyone ever develops any sort of interest in this project, I can write full documentation. Otherwise, "ghost keys".

Fixed a bug causing tau nodes and label keys to have null ids

Changed some clone methods to remove redundant if statements

Added TODO

Signed-off-by: GMX <pbrowning7@gmail.com>