Hopefully address https://github.com/peterbro1/IRDC-CCSK/issues/17
(I'm not sure if this commit message automatically closes this issue, if it does I'll reopen it until we can verify it behaves correctly.)

The solution I came up with is to give the labels themselves a boolean property 'isRestricted'. When a label L is restricted on L', L.isRestricted will be true. The isRestricted property is ignored up until the very top level process returns the final set of actionable labels, where the restricted labels will just be removed

Signed-off-by: GMX <pbrowning7@gmail.com>